Added in port bitwise defines
This will simplify the selection of ports for the shutoff variables.
Currently you have this:

ReefAngel.WaterChangePorts = B10011000;

and now you will be able to do this:

ReefAngel.WaterChangePorts = Port8Bit | Port4Bit | Port5Bit;

It's a little longer to write it out but it makes it much easier to read.
Contribution by dedvalson from forum.