Added alarm clearing for ATO & Overheat to wifi
The commands are as follows:
/mt - clear ATO alarm (/ma was already used)
/mo - clear Overheat alarm

The response will be <MODE>OK</MODE>
I chose to use <MODE> because it already existed and was a little logical.
The response of MODE prompted the use of the /m prefix for the commands.
MA would have been used but that was already for all the memory locations.