Fixed Relay.Write() calls for Feeding / Water change modes
Relay.Write() was being called only before the modes were exited.  This did not allow the controller to process the ports properly during the normal operation mode.  So if water change mode was left on, the ports would stay in the same state that they were before the water change mode was activated (except the ports that were toggled off to enter the mode).